refactor: Move `iroh-router` & examples into `iroh-net`, remove `iroh`
