+++ iroh-net/src/lib.rs
@@ -241,6 +241,7 @@ pub mod dns;
 pub mod endpoint;
 mod magicsock;
 pub mod metrics;
+pub mod protocol;
 pub mod ticket;
 pub mod tls;
