--- iroh-router/Cargo.toml
@@ -1,41 +1,0 @@
-[package]
-name = "iroh-router"
-version = "0.28.0"
-edition = "2021"
-readme = "README.md"
-description = "protocol router support for iroh"
-license = "MIT OR Apache-2.0"
-authors = ["dignifiedquire <me@dignifiedquire.com>", "n0 team"]
-repository = "https://github.com/n0-computer/iroh"
-keywords = ["quic", "networking", "holepunching", "p2p"]
-
-
-[dependencies]
-anyhow = "1.0.91"
-futures-buffered = "0.2.9"
-futures-lite = "2.5"
-futures-util = "0.3.31"
-iroh-net = { version = "0.28.1", path = "../iroh-net" }
-tokio = "1"
-tokio-util = "0.7"
-tracing = "0.1"
-
-# Examples
-clap = { version = "4", features = ["derive"], optional = true }
-tracing-subscriber = { version = "0.3", features = ["env-filter"], optional = true }
-
-[lints]
-workspace = true
-
-
-[features]
-default = []
-examples = ["dep:clap", "dep:tracing-subscriber"]
-
-[[example]]
-name = "search"
-required-features = ["examples"]
-
-[[example]]
-name = "echo"
-required-features = ["examples"]
