--- iroh-router/src/lib.rs
@@ -1,5 +1,0 @@
-mod protocol;
-mod router;
-
-pub use protocol::{ProtocolHandler, ProtocolMap};
-pub use router::{Router, RouterBuilder};
