--- iroh/Cargo.toml
@@ -1,59 +1,0 @@
-[package]
-name = "iroh"
-version = "0.28.1"
-edition = "2021"
-readme = "README.md"
-description = "A toolkit for building distributed applications"
-license = "MIT OR Apache-2.0"
-authors = ["dignifiedquire <me@dignifiedquire.com>", "n0 team"]
-repository = "https://github.com/n0-computer/iroh"
-keywords = ["networking", "p2p", "holepunching", "ipfs"]
-
-# Sadly this also needs to be updated in .github/workflows/ci.yml
-rust-version = "1.76"
-
-[lints]
-workspace = true
-
-[dependencies]
-iroh-base = { version = "0.28.0", features = ["key"] }
-iroh-net = { version = "0.28.1", default-features = false }
-iroh-router = { version = "0.28.0" }
-
-# Examples
-anyhow = { version = "1", optional = true }
-clap = { version = "4", features = ["derive"], optional = true }
-indicatif = { version = "0.17", features = ["tokio"], optional = true }
-parse-size = { version = "=1.0.0", optional = true } # pinned version to avoid bumping msrv to 1.81
-tokio = { version = "1", features = ["full"], optional = true }
-tracing-subscriber = { version = "0.3", features = ["env-filter"], optional = true }
-futures-lite = { version = "2.5", optional = true }
-tracing = { version = "0.1", optional = true }
-bytes = { version = "1.8", optional = true }
-
-
-[features]
-default = ["metrics", "discovery-pkarr-dht"]
-metrics = ["iroh-net/metrics"]
-discovery-local-network = ["iroh-net/discovery-local-network"]
-discovery-pkarr-dht = ["iroh-net/discovery-pkarr-dht"]
-
-examples = [
-  "dep:anyhow",
-  "dep:clap",
-  "dep:indicatif",
-  "dep:parse-size",
-  "dep:tokio",
-  "dep:tracing-subscriber",
-  "dep:futures-lite",
-  "dep:tracing",
-  "dep:bytes",
-]
-
-[package.metadata.docs.rs]
-all-features = true
-rustdoc-args = ["--cfg", "iroh_docsrs"]
-
-[[example]]
-name = "transfer"
-required-features = ["examples"]
